start interpreter code
Also change representation of functions to be one parameter only, and
not a vec of params. We required this already in the syntax, so this is
purely an internal change. The reason for having a vec was efficiency
(less functions when treated as a multi-param function). However, we
need to be able to represent things like

(λx.λy.y) z

which should evaluate to

λy.y

but which would be an application of two-parameter function with only
one argument. Keeping the parameter vec would require building in
currying. Instead we just treat each lambda as single-parameter.